Module starry::loader

Loader for loading apps.

It will read and parse ELF files.

Now these apps are loaded into memory as a part of the kernel image.

Structs

ELFInfo

The information of a given ELF file

ELFSegment

The segment of the elf file, which is used to map the elf file to the memory space

Functions

_app_count 🔒 ⚠

get_app_count 🔒

Get the number of apps.

get_app_data 🔒

Get the data of an app by a given app ID.

get_app_data_by_name 🔒

Get the data of an app by the given app name.

get_app_name 🔒

Get the name of an app by a given app ID.

list_apps 🔒

List all apps.

load_elf 🔒

Load the ELF files by the given app name and return the segments of the ELF file