starry
Module mm

Structs

ACCESSING_USER_MEM_WRAPPER 🔒

Wrapper struct for the per-CPU data ACCESSING_USER_MEM

Statics

ACCESSING_USER_MEM 🔒

_LINKME_ELEMENT_handle_page_fault 🔒

__PERCPU_ACCESSING_USER_MEM 🔒

Functions

access_user_memory

Enables scoped access into user memory, allowing page faults to occur inside kernel.

handle_page_fault 🔒

load_user_app

Load the user app to the user address space.

map_elf 🔒

Map the elf file to the user address space.