starry::ptr
Trait PtrWrapper

pub trait PtrWrapper<T>: Sized {
    type Ptr;

    const ACCESS_FLAGS: MappingFlags;

    // Required methods
    unsafe fn into_inner(self) -> Self::Ptr;
    fn address(&self) -> VirtAddr;

    // Provided methods
    fn get(self) -> LinuxResult<Self::Ptr> { ... }
    fn get_as(self, layout: Layout) -> LinuxResult<Self::Ptr> { ... }
    fn get_as_bytes(self, size: usize) -> LinuxResult<Self::Ptr> { ... }
    fn get_as_array(self, len: usize) -> LinuxResult<Self::Ptr> { ... }
    fn nullable<R>(
        self,
        f: impl FnOnce(Self) -> LinuxResult<R>,
    ) -> LinuxResult<Option<R>> { ... }
}

A trait representing a pointer in user space, which can be converted to a pointer in kernel space through a series of checks.

Converting a PtrWrapper<T> to *T is done by PtrWrapper::get (or get_as_*). It checks whether the pointer along with its layout is valid in the current task’s address space, and raises EFAULT if not.

Required Associated Constants

const ACCESS_FLAGS: MappingFlags

Required Associated Types

type Ptr

Required Methods

unsafe fn into_inner(self) -> Self::Ptr

Unwrap the pointer to the inner type.

This function is unsafe because it assumes that the pointer is valid and points to a valid memory region.

fn address(&self) -> VirtAddr

Get the address of the pointer.

Provided Methods

fn get(self) -> LinuxResult<Self::Ptr>

Get the pointer as a raw pointer to T.

fn get_as(self, layout: Layout) -> LinuxResult<Self::Ptr>

Get the pointer as a raw pointer to T, validating the memory region given by the layout.

fn get_as_bytes(self, size: usize) -> LinuxResult<Self::Ptr>

Get the pointer as a raw pointer to T, validating the memory region specified by the size.

fn get_as_array(self, len: usize) -> LinuxResult<Self::Ptr>

Get the pointer as a raw pointer to T, validating the memory region given by the layout of [T; len].

fn nullable<R>( self, f: impl FnOnce(Self) -> LinuxResult<R>, ) -> LinuxResult<Option<R>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.